fixing build errors
