+++ packages/webapp/src/delegates/api/eden-contract.ts
@@ -1,11 +1,4 @@
-import { devUseFixtureData } from "config";
-import { ElectionState } from "elections/interfaces";
-import {
-    CONTRACT_CURRENT_ELECTION_TABLE,
-    CONTRACT_ELECTION_STATE_TABLE,
-    getTableRows,
-} from "_app";
-import { fixtureElectionState } from "./fixtures";
+import { getElectionState } from "elections/api";
 
 const getDelegates = () => {
     return {}; // TODO
@@ -19,30 +12,6 @@ export const getHeadDelegate = async (): Promise<string | undefined> => {
 export const getChiefDelegates = async (): Promise<string[] | undefined> => {
     const electionState = await getElectionState();
     return electionState?.board;
-};
-
-export const getCurrentElection = async () => {
-    const rows = await getTableRows<any>(CONTRACT_CURRENT_ELECTION_TABLE);
-
-    if (!rows.length) {
-        return undefined;
-    }
-
-    return rows[0];
-};
-
-export const getElectionState = async () => {
-    if (devUseFixtureData) return fixtureElectionState;
-
-    const rows = await getTableRows<ElectionState>(
-        CONTRACT_ELECTION_STATE_TABLE
-    );
-
-    if (!rows.length) {
-        return undefined;
-    }
-
-    return rows[0];
 };
 
 const getMemberBudgetBalance = () => {
