+++ packages/webapp/src/delegates/api/fixtures.ts
@@ -1,13 +1,4 @@
-import { CurrentElection, ElectionState } from "elections/interfaces";
 import { MemberData, MemberStats } from "members";
-
-export const fixtureElectionState: ElectionState = {
-    lead_representative: "edenmember11",
-    board: ["egeon.edev", "edenmember12", "edenmember13", "pip.edev"],
-    last_election_time: "2022-01-16T16:00:00.000",
-};
-
-export const fixtureCurrentElection: CurrentElection = {}; // representing pending state
 
 export const fixtureMemberStats: MemberStats = {
     active_members: 4,
