+++ packages/webapp/src/members/interfaces.ts
@@ -10,10 +10,10 @@ export interface MemberData {
     attributions: string;
     socialHandles: EdenNftSocialHandles;
     inductionVideo: string;
-    status: number;
-    election_participation_status: number;
-    election_rank: number;
-    representative: string;
+    status?: number; // TODO: remove these question marks if possible
+    election_participation_status?: number;
+    election_rank?: number;
+    representative?: string;
     createdAt: number;
     auctionData?: MemberAuctionData;
     assetData?: AssetData;
